Consolidate pubic Tag functions
`process_plugins_data` was moved to `/plugins/openlibrary/tags.py`,
with the other public Tag functions.  Calls to `process_plugins_data`
were failing before this changes.
